+++ packages/components/src/spinner/styles/spinner-styles.js
@@ -7,7 +7,7 @@ import { keyframes } from '@emotion/core';
 /**
  * Internal dependencies
  */
-import { COLORS, config } from '../../utils';
+import { COLORS, CONFIG } from '../../utils';
 
 const spinAnimation = keyframes`
 	from {
@@ -19,15 +19,13 @@ const spinAnimation = keyframes`
 	}
 `;
 
-const topLeft = `calc( ( ${ config( 'spinnerSize' ) } - ${ config(
-	'spinnerSize'
-) } * ( 2 / 3 ) ) / 2 )`;
+const topLeft = `calc( ( ${ CONFIG.spinnerSize } - ${ CONFIG.spinnerSize } * ( 2 / 3 ) ) / 2 )`;
 
 export const StyledSpinner = styled.span`
 	display: inline-block;
 	background-color: ${ COLORS.gray[ 600 ] };
-	width: ${ config( 'spinnerSize' ) };
-	height: ${ config( 'spinnerSize' ) };
+	width: ${ CONFIG.spinnerSize };
+	height: ${ CONFIG.spinnerSize };
 	opacity: 0.7;
 	margin: 5px 11px 0;
 	border-radius: 100%;
@@ -39,11 +37,11 @@ export const StyledSpinner = styled.span`
 		background-color: ${ COLORS.white };
 		top: ${ topLeft };
 		left: ${ topLeft };
-		width: calc( ${ config( 'spinnerSize' ) } / 4.5 );
-		height: calc( ${ config( 'spinnerSize' ) } / 4.5 );
+		width: calc( ${ CONFIG.spinnerSize } / 4.5 );
+		height: calc( ${ CONFIG.spinnerSize } / 4.5 );
 		border-radius: 100%;
-		transform-origin: calc( ${ config( 'spinnerSize' ) } / 3 )
-			calc( ${ config( 'spinnerSize' ) } / 3 );
+		transform-origin: calc( ${ CONFIG.spinnerSize } / 3 )
+			calc( ${ CONFIG.spinnerSize } / 3 );
 		animation: ${ spinAnimation } 1s infinite linear;
 	}
 `;