+++ packages/rich-text/src/component/index.js
@@ -1,7 +1,7 @@
 /**
  * WordPress dependencies
  */
-import { useRef, useState, useLayoutEffect } from '@wordpress/element';
+import { useRef, useLayoutEffect, useReducer } from '@wordpress/element';
 import { useMergeRefs, useRefEffect } from '@wordpress/compose';
 
 /**
@@ -20,82 +20,24 @@ import { useIndentListItemOnSpace } from './use-indent-list-item-on-space';
 import { useInputAndSelection } from './use-input-and-selection';
 import { useDelete } from './use-delete';
 
-/** @typedef {import('@wordpress/element').WPSyntheticEvent} WPSyntheticEvent */
-
 export function useRichText( {
 	value = '',
 	selectionStart,
 	selectionEnd,
 	placeholder,
 	preserveWhiteSpace,
-	format = 'string',
 	onSelectionChange,
 	onChange,
 	__unstableMultilineTag: multilineTag,
 	__unstableDisableFormats: disableFormats,
-	__unstableOnCreateUndoLevel: onCreateUndoLevel,
 	__unstableIsSelected: isSelected,
 	__unstableDependencies,
 	__unstableAfterParse,
 	__unstableBeforeSerialize,
 	__unstableAddInvisibleFormats,
 } ) {
+	const [ , forceRender ] = useReducer( () => ( {} ) );
 	const ref = useRef();
-	const [ activeFormats = [], setActiveFormats ] = useState();
-
-	/**
-	 * Converts the outside data structure to our internal representation.
-	 *
-	 * @param {*} string The outside value, data type depends on props.
-	 *
-	 * @return {Object} An internal rich-text value.
-	 */
-	function formatToValue( string ) {
-		if ( disableFormats ) {
-			return {
-				text: string,
-				formats: Array( string.length ),
-				replacements: Array( string.length ),
-			};
-		}
-
-		if ( format !== 'string' ) {
-			return string;
-		}
-
-		const result = create( {
-			html: string,
-			multilineTag,
-			multilineWrapperTags:
-				multilineTag === 'li' ? [ 'ul', 'ol' ] : undefined,
-			preserveWhiteSpace,
-		} );
-
-		result.formats = __unstableAfterParse( result );
-
-		return result;
-	}
-
-	/**
-	 * Converts the internal value to the external data format.
-	 *
-	 * @param {Object} val The internal rich-text value.
-	 *
-	 * @return {*} The external data format, data type depends on props.
-	 */
-	function valueToFormat( val ) {
-		if ( disableFormats ) {
-			return val.text;
-		}
-
-		if ( format !== 'string' ) {
-			return;
-		}
-
-		val.formats = __unstableBeforeSerialize( val );
-
-		return toHTMLString( { value: val, multilineTag, preserveWhiteSpace } );
-	}
 
 	function createRecord() {
 		const {
@@ -132,55 +74,66 @@ export function useRichText( {
 	// Internal values are updated synchronously, unlike props and state.
 	const _value = useRef( value );
 	const record = useRef();
-	const lastHistoryValue = useRef( value );
 
 	function setRecordFromProps() {
-		record.current = formatToValue( value );
+		_value.current = value;
+		record.current = create( {
+			html: value,
+			multilineTag,
+			multilineWrapperTags:
+				multilineTag === 'li' ? [ 'ul', 'ol' ] : undefined,
+			preserveWhiteSpace,
+		} );
+		if ( disableFormats ) {
+			record.current.formats = Array( value.length );
+			record.current.replacements = Array( value.length );
+		}
+		record.current.formats = __unstableAfterParse( record.current );
 		record.current.start = selectionStart;
 		record.current.end = selectionEnd;
 	}
 
+	const hadSelectionUpdate = useRef( false );
+
 	if ( ! record.current ) {
 		setRecordFromProps();
-	}
-
-	function createUndoLevel() {
-		// If the content is the same, no level needs to be created.
-		if ( lastHistoryValue.current === _value.current ) {
-			return;
-		}
-
-		onCreateUndoLevel();
-		lastHistoryValue.current = _value.current;
+	} else if (
+		selectionStart !== record.current.start ||
+		selectionEnd !== record.current.end
+	) {
+		hadSelectionUpdate.current = isSelected;
+		record.current = {
+			...record.current,
+			start: selectionStart,
+			end: selectionEnd,
+		};
 	}
 
 	/**
 	 * Sync the value to global state. The node tree and selection will also be
 	 * updated if differences are found.
 	 *
-	 * @param {Object}  newRecord         The record to sync and apply.
-	 * @param {Object}  $2                Named options.
-	 * @param {boolean} $2.withoutHistory If true, no undo level will be
-	 *                                    created.
+	 * @param {Object} newRecord The record to sync and apply.
 	 */
-	function handleChange( newRecord, { withoutHistory } = {} ) {
-		if ( disableFormats ) {
-			newRecord.formats = Array( newRecord.text.length );
-			newRecord.replacements = Array( newRecord.text.length );
-		}
-
+	function handleChange( newRecord ) {
 		applyRecord( newRecord );
 
-		const {
-			start,
-			end,
-			formats,
-			text,
-			activeFormats: newActiveFormats = [],
-		} = newRecord;
+		if ( disableFormats ) {
+			_value.current = newRecord.text;
+		} else {
+			_value.current = toHTMLString( {
+				value: {
+					...newRecord,
+					formats: __unstableBeforeSerialize( newRecord ),
+				},
+				multilineTag,
+				preserveWhiteSpace,
+			} );
+		}
 
-		_value.current = valueToFormat( newRecord );
 		record.current = newRecord;
+
+		const { start, end, formats, text } = newRecord;
 
 		// Selection must be updated first, so it is recorded in history when
 		// the content change happens.
@@ -189,15 +142,10 @@ export function useRichText( {
 			__unstableFormats: formats,
 			__unstableText: text,
 		} );
-		setActiveFormats( newActiveFormats );
-
-		if ( ! withoutHistory ) {
-			createUndoLevel();
-		}
+		forceRender();
 	}
 
 	function applyFromProps( { domOnly } = {} ) {
-		_value.current = value;
 		setRecordFromProps();
 		applyRecord( record.current, { domOnly } );
 	}
@@ -213,24 +161,13 @@ export function useRichText( {
 
 	// Value updates must happen synchonously to avoid overwriting newer values.
 	useLayoutEffect( () => {
-		if ( ! didMount.current ) {
+		if ( ! hadSelectionUpdate.current ) {
 			return;
 		}
 
-		if (
-			isSelected &&
-			( selectionStart !== record.current.start ||
-				selectionEnd !== record.current.end )
-		) {
-			applyFromProps();
-		} else {
-			record.current = {
-				...record.current,
-				start: selectionStart,
-				end: selectionEnd,
-			};
-		}
-	}, [ selectionStart, selectionEnd, isSelected ] );
+		applyFromProps();
+		hadSelectionUpdate.current = false;
+	}, [ hadSelectionUpdate.current ] );
 
 	function focus() {
 		ref.current.focus();
@@ -240,11 +177,11 @@ export function useRichText( {
 	const mergedRefs = useMergeRefs( [
 		ref,
 		useDefaultStyle(),
-		useBoundaryStyle( { activeFormats } ),
+		useBoundaryStyle( { record } ),
 		useInlineWarning(),
 		useCopyHandler( { record, multilineTag, preserveWhiteSpace } ),
 		useSelectObject(),
-		useFormatBoundaries( { record, applyRecord, setActiveFormats } ),
+		useFormatBoundaries( { record, applyRecord } ),
 		useDelete( {
 			createRecord,
 			handleChange,
@@ -260,10 +197,8 @@ export function useRichText( {
 			applyRecord,
 			createRecord,
 			handleChange,
-			createUndoLevel,
 			isSelected,
 			onSelectionChange,
-			setActiveFormats,
 		} ),
 		useRefEffect( () => {
 			if ( didMount.current ) {
@@ -281,7 +216,6 @@ export function useRichText( {
 		onChange: handleChange,
 		onFocus: focus,
 		ref: mergedRefs,
-		hasActiveFormats: activeFormats.length,
 	};
 }