+++ packages/rich-text/src/component/use-input-and-selection.js
@@ -63,7 +63,6 @@ export function useInputAndSelection( props ) {
 
 		let isComposing = false;
 		let rafId;
-		let timeout;
 
 		function onInput( event ) {
 			// Do not trigger a change if characters are being composed.
@@ -81,16 +80,11 @@ export function useInputAndSelection( props ) {
 				inputType = event.inputType;
 			}
 
-			if ( ! inputType && event && event.nativeEvent ) {
-				inputType = event.nativeEvent.inputType;
-			}
-
 			const {
 				record,
 				applyRecord,
 				createRecord,
 				handleChange,
-				createUndoLevel,
 			} = propsRef.current;
 
 			// The browser formatted something or tried to insert HTML.
@@ -119,11 +113,7 @@ export function useInputAndSelection( props ) {
 				formats: oldActiveFormats,
 			} );
 
-			handleChange( change, { withoutHistory: true } );
-
-			// Create an undo level when input stops for over a second.
-			defaultView.clearTimeout( timeout );
-			timeout = defaultView.setTimeout( createUndoLevel, 1000 );
+			handleChange( change );
 		}
 
 		/**
@@ -144,7 +134,6 @@ export function useInputAndSelection( props ) {
 				createRecord,
 				isSelected,
 				onSelectionChange,
-				setActiveFormats,
 			} = propsRef.current;
 
 			if ( event.type !== 'selectionchange' && ! isSelected ) {
@@ -209,7 +198,6 @@ export function useInputAndSelection( props ) {
 			record.current = newValue;
 			applyRecord( newValue, { domOnly: true } );
 			onSelectionChange( start, end );
-			setActiveFormats( newActiveFormats );
 		}
 
 		function onCompositionStart() {
@@ -236,12 +224,7 @@ export function useInputAndSelection( props ) {
 		}
 
 		function onFocus() {
-			const {
-				record,
-				isSelected,
-				onSelectionChange,
-				setActiveFormats,
-			} = propsRef.current;
+			const { record, isSelected, onSelectionChange } = propsRef.current;
 
 			if ( ! isSelected ) {
 				// We know for certain that on focus, the old selection is invalid.
@@ -256,7 +239,6 @@ export function useInputAndSelection( props ) {
 					activeFormats: EMPTY_ACTIVE_FORMATS,
 				};
 				onSelectionChange( index, index );
-				setActiveFormats( EMPTY_ACTIVE_FORMATS );
 			} else {
 				onSelectionChange( record.current.start, record.current.end );
 			}
@@ -309,7 +291,6 @@ export function useInputAndSelection( props ) {
 				handleSelectionChange
 			);
 			defaultView.cancelAnimationFrame( rafId );
-			defaultView.clearTimeout( timeout );
 		};
 	}, [] );
 }